Removed dependency from services to base Rails models #7125
